refactor: organized routes into routers
