+++ package.json
@@ -3,9 +3,9 @@
 	"version": "1.0.0",
 	"description": "",
 	"type": "module",
-	"main": "index.js",
+	"main": "app.js",
 	"scripts": {
-		"dev": "nodemon src/index.js"
+		"dev": "nodemon src/app.js"
 	},
 	"repository": {
 		"type": "git",
